+++ app/src/main/java/com/artem/coffeeshop/presentation/account/CreateAccountFragment.kt
@@ -7,27 +7,26 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.lifecycle.ViewModelProvider
 import com.artem.coffeeshop.R
-import com.artem.coffeeshop.presentation.account.ViewModelAccount.ViewModelAccount
+import com.artem.coffeeshop.databinding.FragmentCreateAccountBinding
+import com.artem.coffeeshop.databinding.FragmentCreateAccountBinding.inflate
+import com.artem.coffeeshop.presentation.account.viewModelAccount.ViewModelAccount
 
 
 
 class  CreateAccountFragment : Fragment() {
 
 
+
     override fun onCreateView(inflater: LayoutInflater,
                               container: ViewGroup?,
                               savedInstanceState: Bundle?
     ): View? {
-
+        
         return inflater.inflate(R.layout.fragment_create_account, container, false)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
-        val model = ViewModelProvider(requireActivity()).get(ViewModelAccount::class.java)
-
-
 
 
     }
