--- app/src/main/java/com/artem/coffeeshop/presentation/account/ViewModelAccount/EnterAccountFragment.kt
@@ -1,36 +1,0 @@
-package com.artem.coffeeshop.presentation.account.ViewModelAccount
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.navigation.findNavController
-import com.artem.coffeeshop.R
-
-
-
-class EnterAccountFragment : Fragment() {
-
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-
-
-        return inflater.inflate(R.layout.fragment_enter_account, container, false)
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-
-
-
-
-        view.findNavController().navigate(R.id.action_enterAccountFragment_to_createAccountFragment)
-
-    }
-}
