+++ app/src/main/java/com/artem/coffeeshop/presentation/account/viewModelAccount/ViewModelAccount.kt
@@ -1,4 +1,4 @@
-package com.artem.coffeeshop.presentation.account.ViewModelAccount
+package com.artem.coffeeshop.presentation.account.viewModelAccount
 
 
 import androidx.lifecycle.ViewModel
