+++ app/src/main/res/layout/fragment_enter_account.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".presentation.account.ViewModelAccount.EnterAccountFragment">
+    tools:context=".presentation.account.viewModelAccount.EnterAccountFragment">
 
  <LinearLayout
      android:layout_width="match_parent"
